Core: split the image decoder by format
One module per container (png with apng, jpeg, webp, gif), each carrying
its own feature gates and compiled-out stubs, behind the shared sniff,
budgets and frame model in mod.rs. Callers keep their paths through the
re-exports.
